# Simplifies readFrom and WriteTo, and improves their performance.