@omerfirmak 2023-10-12
> @lemire you are trading space for time

`-benchmem` to see the GC hit you are taking. We, at Nethermind, don't depend on `bitset` anymore (due to https://github.com/bits-and-blooms/bitset/issues/134).