@lemire 2023-10-12
@thanhpk 

> it would alloc a big chunk of memory, crashing some programs if the bitset is big enough

I don't think it would. The overhead memory usage should still effectively constant: 

```Go

func readUint64Array(reader io.Reader, data []uint64) error {
	length := len(data)
	for i := 0; i < length; i += 512 {
		end := i + 512
		if end > length {
			end = length
		}
		chunk := data[i:end]
		err := binary.Read(reader, binaryOrder, chunk)
		if err != nil {
			return err
		}
	}
	return nil
}

func writeUint64Array(writer io.Writer, data []uint64) error {
	for i := 0; i < len(data); i += 512 {
		end := i + 512
		if end > len(data) {
			end = len(data)
		}
		chunk := data[i:end]
		err := binary.Write(writer, binaryOrder, chunk)
		if err != nil {
			return err
		}
	}
	return nil
}
```

@omerfirmak 

>  I would rather keep the GC at minimum.

Yeah. It seems that `binary.Read` and `binary.Write` are allocating functions. Hmmm...